Add form validation and various changes
Added form validation for signup and login pages. Converted post and
comment usernames to profile page links. Added hover overlay to profile
picture. Changed font-family to Lato for entire site and Titan One for
logo.

+++ src/components/ImageUploadPage.jsx
@@ -145,8 +145,9 @@ function ImageUploadPage() {
   async function uploadFile() {
     if(profilePicture) {
       try {
-        await deleteObject(ref(storage, user.photoURL));
-
+        if(user.photoURL !== 'default/default-profile-picture.png') {
+          await deleteObject(ref(storage, user.photoURL));
+        }
         const storageURL = `${user.uid}/ProfilePicture/${file.name}`;
         const storageRef = ref(storage, storageURL);
         const data = await fetch(croppedImage);
